pyproject.toml
[project]
name = "smolalbert"
version = "0.1.0"
description = "my personal deep research agent"
authors = [
    {name = "blackccpie",email = "16404298+blackccpie@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio (>=5.42.0,<6.0.0)",
    "smolagents (>=1.21.1,<2.0.0)",
    "tavily-python (>=0.7.10,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
